tests/wpt/web-platform-tests/font-access/font_access-chooser-multiple.tentative.manual.https.html

<!doctype html>
<title>Local Font Access: Multiple choosers</title>
<meta charset=utf-8>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script>
  promise_test(async t => {
    await new Promise(resolve => {
      window.addEventListener('DOMContentLoaded', resolve);
    });
    // Small delay to give chrome's test automation a chance to actually install
    // itself.
    await new Promise(resolve => step_timeout(resolve, 100));

    await window.test_driver.bless('show a font chooser.<br>Please select at least one font.');
    const promise = navigator.fonts.query()
    promise_rejects_dom(t, 'SecurityError', navigator.fonts.query());
    const fonts = await promise;
  }, 'query() multiple choosers');
</script>
